requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
types-pytz==2021.3.4
types-PyYAML==6.0.3
typing-extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
